requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.11
Django==4.0.8
djangorestframework-simplejwt==5.2.2
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
idna==3.3
Pillow==9.3.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
reportlab==3.6.11
requests==2.27.1
six==1.16.0
sslcommerz-lib==1.0
urllib3==1.26.8
virtualenv==20.14.1
whitenoise==6.2.0
widgetsnbextension==3.5.2
xhtml2pdf==0.2.8